requirements.txt
chainlit==2.4.1
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.0.0
pyyaml>=6.0.1
pydantic>=2.7.2,<3